refactor: update Button and Input components with new classes and structure; adjust global styles in tailwind.css for consistency
