+++ @/components/input.tsx
@@ -1,22 +1,100 @@
 import * as React from 'react';
-
 import {cn} from '@/lib/utils';
+import { cva, VariantProps } from 'class-variance-authority';
+import { useState } from 'react';
+import { IconClose } from '~/components/Icon';
+import clsx from 'clsx';
+
+const inputVariants = cva(
+  'w-full !shadow-none focus:ring-0 focus-visible:outline-none placeholder:text-text-primary bg-transparent',
+  {
+    variants: {
+      variant: {
+        default: 'flex w-full rounded border-2 border-border-subtle px-3 py-3 text-sm file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:border-border hover:border-border disabled:cursor-not-allowed disabled:opacity-50',
+        search:
+          'px-0 py-2 text-sm placeholder-foreground-subtle w-full focus:ring-0 border-x-0 border-t-0 transition border-b-2 border-bar/10 focus:border-bar/50',
+        minisearch:
+          'hidden md:inline-block text-left lg:text-right border-b transition border-transparent -mb-px border-x-0 border-t-0 appearance-none px-0 py-1 focus:ring-transparent placeholder:opacity-20 placeholder:text-inherit focus:border-bar/50',
+        error: 'border-red-500',
+        custom: '',
+      },
+    },
+    defaultVariants: {
+      variant: 'default',
+    },
+  },
+);
 
 export interface InputProps
-  extends React.InputHTMLAttributes<HTMLInputElement> {}
+  extends React.InputHTMLAttributes<HTMLInputElement>, VariantProps<typeof inputVariants> {
+    suffix?: React.ReactNode;
+    prefixElement?: React.ReactNode;
+    onClear?: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  }
 
 const Input = React.forwardRef<HTMLInputElement, InputProps>(
-  ({className, type, ...props}, ref) => {
-    return (
+  ({className,
+    type,
+    variant = 'default',
+    prefixElement,
+    suffix,
+    onFocus,
+    onBlur,
+    onClear,
+    ...rest}, ref) => {
+      let [focused, setFocused] = useState(false);
+    let commonClasses = cn(
+      'w-full rounded-sm border px-3 py-2.5',
+      focused ? 'border-border' : 'border-border-subtle',
+      className,
+    );
+
+    let handleClear = (e: any) => {
+      e.preventDefault();
+      e.stopPropagation();
+      e.currentTarget.previousSibling.value = '';
+      if (onClear) onClear(e);
+    };
+    if (type === 'search') {
+      suffix = <IconClose className="cursor-pointer" onClick={handleClear} />;
+    }
+    let hasChild = Boolean(prefixElement || suffix);
+
+    let rawInput = (
       <input
         type={type}
-        className={cn(
-          'flex w-full rounded border-2 border-bar-subtle px-3 py-3 text-sm file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:border-bar hover:border-bar disabled:cursor-not-allowed disabled:opacity-50 placeholder:text-foreground-subtle',
-          className,
-        )}
         ref={ref}
-        {...props}
+        className={cn(
+          hasChild
+            ? 'relatvie grow border-none p-0'
+            : commonClasses,
+          inputVariants({variant}),
+        )}
+        onFocus={(e) => {
+          setFocused(true);
+          if (onFocus) onFocus(e);
+        }}
+        onBlur={(e) => {
+          setFocused(false);
+          if (onBlur) onBlur(e);
+        }}
+        {...rest}
       />
+    );
+
+    return hasChild ? (
+      <div
+        className={clsx(
+          commonClasses,
+          'flex items-center gap-2 overflow-hidden border p-2.5',
+        )}
+      >
+        {prefixElement}
+        {rawInput}
+        {suffix}
+      </div>
+    ) : (
+      rawInput
     );
   },
 );
